feat: remove deprecated resource-available-form
